+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Bug fix: calculate off-chain balance correctly
 
+- Docker: update c-lightning and Bitcoin Core
+
 ## 0.2.0-rc.1 - 2018-12-03
 
 - Channel management GUI